Move selectors to dedicated folders
