+++ ts/features/fci/screens/valid/FciQtspClausesScreen.tsx
@@ -14,7 +14,7 @@ import { useIONavigation } from "../../../../navigation/params/AppParamsList";
 import { useIODispatch, useIOSelector } from "../../../../store/hooks";
 import { emptyContextualHelp } from "../../../../utils/emptyContextualHelp";
 import { loadServicePreference } from "../../../services/details/store/actions/preference";
-import { servicePreferencePotByIdSelector } from "../../../services/details/store/reducers";
+import { servicePreferencePotByIdSelector } from "../../../services/details/store/selectors";
 import { isServicePreferenceResponseSuccess } from "../../../services/details/types/ServicePreferenceResponse";
 import { trackFciUxConversion } from "../../analytics";
 import GenericErrorComponent from "../../components/GenericErrorComponent";
