+++ ts/features/fims/singleSignOn/saga/handleFimsAuthorizationOrImplicitCodeFlow.ts
@@ -26,7 +26,7 @@ import {
   lollipopPublicKeySelector
 } from "../../../lollipop/store/reducers/lollipop";
 import { lollipopRequestInit } from "../../../lollipop/utils/fetch";
-import { serviceDetailsByIdSelector } from "../../../services/details/store/reducers";
+import { serviceDetailsByIdSelector } from "../../../services/details/store/selectors";
 import {
   fimsCtaTextSelector,
   fimsEphemeralSessionOniOSSelector,
