+++ ts/features/messages/components/MessageDetail/MessageDetailsFooter.tsx
@@ -9,7 +9,7 @@ import I18n from "i18next";
 import { ServiceId } from "../../../../../definitions/backend/ServiceId";
 import { useIOSelector } from "../../../../store/hooks";
 import { formatPaymentNoticeNumber } from "../../../payments/common/utils";
-import { serviceMetadataByIdSelector } from "../../../services/details/store/reducers";
+import { serviceMetadataByIdSelector } from "../../../services/details/store/selectors";
 import { ContactsListItem } from "./ContactsListItem";
 import {
   ShowMoreItems,
