+++ ts/features/messages/saga/handleThirdPartyMessage.ts
@@ -21,7 +21,7 @@ import { SagaCallReturnType } from "../../../types/utils";
 import { unknownToReason } from "../utils";
 import { ThirdPartyMessageWithContent } from "../../../../definitions/backend/ThirdPartyMessageWithContent";
 import { TagEnum } from "../../../../definitions/backend/MessageCategoryPN";
-import { serviceDetailsByIdSelector } from "../../services/details/store/reducers";
+import { serviceDetailsByIdSelector } from "../../services/details/store/selectors";
 import { ServiceDetails } from "../../../../definitions/services/ServiceDetails";
 import { thirdPartyKind } from "../store/reducers/thirdPartyById";
