+++ ts/features/pn/store/sagas/watchPnSaga.ts
@@ -12,7 +12,7 @@ import { isPnTestEnabledSelector } from "../../../../store/reducers/persistedPre
 import { SessionToken } from "../../../../types/SessionToken";
 import { isTestEnv } from "../../../../utils/environment";
 import { loadServicePreference } from "../../../services/details/store/actions/preference";
-import { servicePreferencePotByIdSelector } from "../../../services/details/store/reducers";
+import { servicePreferencePotByIdSelector } from "../../../services/details/store/selectors";
 import { isServicePreferenceResponseSuccess } from "../../../services/details/types/ServicePreferenceResponse";
 import {
   trackPNServiceStatusChangeError,
