+++ ts/features/services/details/components/ServiceDetailsMetadata.tsx
@@ -12,7 +12,7 @@ import I18n from "i18next";
 import { ServiceId } from "../../../../../definitions/services/ServiceId";
 import { ServiceMetadata } from "../../../../../definitions/services/ServiceMetadata";
 import { useIOSelector } from "../../../../store/hooks";
-import { serviceMetadataByIdSelector } from "../store/reducers";
+import { serviceMetadataByIdSelector } from "../store/selectors";
 import { handleItemOnPress } from "../../../../utils/url";
 import * as analytics from "../../common/analytics";
