+++ ts/features/services/details/store/selectors/__tests__/dataById.test.ts
@@ -1,23 +1,16 @@
-import * as pot from "@pagopa/ts-commons/lib/pot";
 import {
   NonEmptyString,
   OrganizationFiscalCode
 } from "@pagopa/ts-commons/lib/strings";
-import { Action, createStore } from "redux";
+import { Action } from "redux";
 import { ServiceId } from "../../../../../../../definitions/services/ServiceId";
 import { ServiceDetails } from "../../../../../../../definitions/services/ServiceDetails";
 import { StandardServiceCategoryEnum } from "../../../../../../../definitions/services/StandardServiceCategory";
 import { ScopeTypeEnum } from "../../../../../../../definitions/services/ScopeType";
 import { SpecialServiceCategoryEnum } from "../../../../../../../definitions/services/SpecialServiceCategory";
-import { applicationChangeState } from "../../../../../../store/actions/application";
-import {
-  logoutSuccess,
-  sessionExpired
-} from "../../../../../authentication/common/store/actions";
 import { loadServiceDetail } from "../../actions/details";
 import { appReducer } from "../../../../../../store/reducers";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { reproduceSequence } from "../../../../../../utils/tests";
 import {
   isErrorServiceDetailsByIdSelector,
   isLoadingServiceDetailsByIdSelector,
@@ -37,71 +30,7 @@ const service = {
   }
 } as ServiceDetails;
 
-describe("serviceById reducer", () => {
-  it("should have initial state", () => {
-    const state = appReducer(undefined, applicationChangeState("active"));
-
-    expect(state.features.services.details.dataById).toStrictEqual({});
-  });
-
-  it("should handle loadServiceDetail action", () => {
-    const state = appReducer(undefined, applicationChangeState("active"));
-    const store = createStore(appReducer, state as any);
-
-    store.dispatch(loadServiceDetail.request(serviceId));
-
-    expect(store.getState().features.services.details.dataById).toStrictEqual({
-      serviceId: pot.noneLoading
-    });
-
-    store.dispatch(loadServiceDetail.success(service));
-    expect(store.getState().features.services.details.dataById).toStrictEqual({
-      serviceId: pot.some(service)
-    });
-
-    const tError = {
-      error: new Error("load failed"),
-      service_id: serviceId
-    };
-
-    store.dispatch(loadServiceDetail.failure(tError));
-    expect(store.getState().features.services.details.dataById).toStrictEqual({
-      serviceId: pot.someError(service, new Error("load failed"))
-    });
-  });
-
-  it("should handle logoutSuccess action", () => {
-    const sequenceOfActions: ReadonlyArray<Action> = [
-      applicationChangeState("active"),
-      loadServiceDetail.success(service),
-      logoutSuccess()
-    ];
-
-    const state: GlobalState = reproduceSequence(
-      {} as GlobalState,
-      appReducer,
-      sequenceOfActions
-    );
-    expect(state.features.services.details.dataById).toEqual({});
-  });
-
-  it("should handle sessionExpired action", () => {
-    const sequenceOfActions: ReadonlyArray<Action> = [
-      applicationChangeState("active"),
-      loadServiceDetail.success(service),
-      sessionExpired()
-    ];
-
-    const state: GlobalState = reproduceSequence(
-      {} as GlobalState,
-      appReducer,
-      sequenceOfActions
-    );
-    expect(state.features.services.details.dataById).toEqual({});
-  });
-});
-
-describe("serviceById selectors", () => {
+describe("dataById selectors", () => {
   describe("serviceByIdSelector", () => {
     it("should return the ServiceDetails when pot.some", () => {
       const serviceById = serviceDetailsByIdSelector(
