+++ ts/features/services/home/store/reducers/__tests__/store.test.ts
@@ -1,24 +1,12 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import MockDate from "mockdate";
-import { Action, createStore } from "redux";
-import {
-  featuredInstitutionsSelector,
-  featuredServicesSelector,
-  isErrorFeaturedInstitutionsSelector,
-  isErrorFeaturedServicesSelector,
-  isLoadingFeaturedInstitutionsSelector,
-  isLoadingFeaturedServicesSelector,
-  paginatedInstitutionsCurrentPageSelector,
-  paginatedInstitutionsLastPageSelector,
-  paginatedInstitutionsSelector
-} from "..";
+import { createStore } from "redux";
 import { FeaturedService } from "../../../../../../../definitions/services/FeaturedService";
 import { FeaturedServices } from "../../../../../../../definitions/services/FeaturedServices";
 import { InstitutionsResource } from "../../../../../../../definitions/services/InstitutionsResource";
 import { OrganizationFiscalCode } from "../../../../../../../definitions/services/OrganizationFiscalCode";
 import { applicationChangeState } from "../../../../../../store/actions/application";
 import { appReducer } from "../../../../../../store/reducers";
-import { GlobalState } from "../../../../../../store/reducers/types";
 import { NetworkError } from "../../../../../../utils/errors";
 import {
   featuredInstitutionsGet,
@@ -211,114 +199,6 @@ describe("Services home paginatedInstitutions reducer", () => {
   });
 });
 
-describe("Services home paginatedInstitutions selectors", () => {
-  describe("paginatedInstitutionsSelector", () => {
-    it("should return the InstitutionsResource when pot.some", () => {
-      const paginatedInstitutions = paginatedInstitutionsSelector(
-        appReducer(
-          {} as GlobalState,
-          paginatedInstitutionsGet.success({
-            institutions: MOCK_INSTITUTIONS,
-            count: 23,
-            offset: 0,
-            limit: 3
-          })
-        )
-      );
-      expect(paginatedInstitutions).toStrictEqual({
-        institutions: MOCK_INSTITUTIONS,
-        count: 23,
-        offset: 0,
-        limit: 3
-      });
-    });
-
-    it("should return undefined when not pot.some", () => {
-      expect(
-        paginatedInstitutionsSelector(appReducer(undefined, {} as Action))
-      ).toBeUndefined();
-
-      expect(
-        paginatedInstitutionsSelector(
-          appReducer(
-            {} as GlobalState,
-            paginatedInstitutionsGet.request({
-              offset: 0,
-              limit: 3
-            })
-          )
-        )
-      ).toBeUndefined();
-
-      expect(
-        paginatedInstitutionsSelector(
-          appReducer(
-            {} as GlobalState,
-            paginatedInstitutionsGet.failure(MOCK_NETWORK_ERROR)
-          )
-        )
-      ).toBeUndefined();
-    });
-  });
-
-  describe("paginatedInstitutionsCurrentPageSelector", () => {
-    [
-      { expectedPage: 0, offset: 0, limit: 20, count: 55 },
-      { expectedPage: 1, offset: 20, limit: 20, count: 55 },
-      { expectedPage: 2, offset: 40, limit: 20, count: 55 },
-      { expectedPage: -1, offset: 60, limit: 20, count: 55 }
-    ].forEach(({ expectedPage, offset, limit, count }) => {
-      it(`should return page "${expectedPage}" when offset is "${offset}", limit is "${limit}" and count is "${count}"`, () => {
-        const currentPage = paginatedInstitutionsCurrentPageSelector(
-          appReducer(
-            {} as GlobalState,
-            paginatedInstitutionsGet.success({
-              institutions: MOCK_INSTITUTIONS,
-              count,
-              offset,
-              limit
-            })
-          )
-        );
-
-        expect(currentPage).toStrictEqual(expectedPage);
-      });
-    });
-
-    it(`should return page "0" when not pot.some`, () => {
-      expect(
-        paginatedInstitutionsCurrentPageSelector(
-          appReducer(undefined, {} as Action)
-        )
-      ).toStrictEqual(0);
-    });
-  });
-
-  describe("paginatedInstitutionsLastPageSelector", () => {
-    [
-      { isLastPage: false, offset: 0, limit: 20, count: 55 },
-      { isLastPage: false, offset: 20, limit: 20, count: 55 },
-      { isLastPage: true, offset: 40, limit: 20, count: 55 }
-    ].forEach(({ isLastPage, offset, limit, count }) => {
-      it(`should return isLastPage===${isLastPage} when offset is "${offset}", limit is "${limit}" and count is "${count}"`, () => {
-        const lastPage = paginatedInstitutionsLastPageSelector(
-          appReducer(
-            {} as GlobalState,
-            paginatedInstitutionsGet.success({
-              institutions: MOCK_INSTITUTIONS,
-              count,
-              offset,
-              limit
-            })
-          )
-        );
-
-        expect(lastPage).toStrictEqual(isLastPage);
-      });
-    });
-  });
-});
-
 describe("Services home featuredInstitutions reducer", () => {
   it("should have initial state", () => {
     const state = appReducer(undefined, applicationChangeState("active"));
@@ -364,100 +244,6 @@ describe("Services home featuredInstitutions reducer", () => {
   });
 });
 
-describe("Services home featuredInstitutions selectors", () => {
-  describe("featuredInstitutionsSelector", () => {
-    it("should return FeaturedInstitutions when pot.some", () => {
-      const featuredInstitutions = featuredInstitutionsSelector(
-        appReducer(
-          {} as GlobalState,
-          featuredInstitutionsGet.success({
-            institutions: MOCK_INSTITUTIONS
-          })
-        )
-      );
-      expect(featuredInstitutions).toStrictEqual(MOCK_INSTITUTIONS);
-    });
-
-    it("should return undefined when not pot.some", () => {
-      expect(
-        featuredInstitutionsSelector(appReducer(undefined, {} as Action))
-      ).toStrictEqual([]);
-
-      expect(
-        featuredInstitutionsSelector(
-          appReducer({} as GlobalState, featuredInstitutionsGet.request())
-        )
-      ).toStrictEqual([]);
-
-      expect(
-        featuredInstitutionsSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredInstitutionsGet.failure(MOCK_NETWORK_ERROR)
-          )
-        )
-      ).toStrictEqual([]);
-    });
-  });
-
-  describe("isLoadingFeaturedInstitutionsSelector", () => {
-    it("should return true when pot.loading", () => {
-      const isLoading = isLoadingFeaturedInstitutionsSelector(
-        appReducer({} as GlobalState, featuredInstitutionsGet.request())
-      );
-      expect(isLoading).toStrictEqual(true);
-    });
-
-    it("should return false when not pot.loading", () => {
-      expect(
-        isLoadingFeaturedInstitutionsSelector(
-          appReducer(undefined, {} as Action)
-        )
-      ).toStrictEqual(false);
-
-      expect(
-        isLoadingFeaturedInstitutionsSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredInstitutionsGet.success({
-              institutions: MOCK_INSTITUTIONS
-            })
-          )
-        )
-      ).toStrictEqual(false);
-    });
-  });
-
-  describe("isErrorFeaturedInstitutionsSelector", () => {
-    it("should return true when pot.error", () => {
-      const isError = isErrorFeaturedInstitutionsSelector(
-        appReducer(
-          {} as GlobalState,
-          featuredInstitutionsGet.failure(MOCK_NETWORK_ERROR)
-        )
-      );
-      expect(isError).toStrictEqual(true);
-    });
-
-    it("should return false when not pot.error", () => {
-      expect(
-        isErrorFeaturedInstitutionsSelector(appReducer(undefined, {} as Action))
-      ).toStrictEqual(false);
-
-      expect(
-        isErrorFeaturedInstitutionsSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredInstitutionsGet.success({
-              institutions: MOCK_INSTITUTIONS
-            })
-          )
-        )
-      ).toStrictEqual(false);
-    });
-  });
-});
-
 describe("Services home featuredServices reducer", () => {
   it("should have initial state", () => {
     const state = appReducer(undefined, applicationChangeState("active"));
@@ -500,97 +286,5 @@ describe("Services home featuredServices reducer", () => {
         MOCK_NETWORK_ERROR
       )
     );
-  });
-});
-
-describe("Services home featuredServices selectors", () => {
-  describe("featuredServicesSelector", () => {
-    it("should return FeaturedServices when pot.some", () => {
-      const featuredServices = featuredServicesSelector(
-        appReducer(
-          {} as GlobalState,
-          featuredServicesGet.success({
-            services: MOCK_FEATURED_SERVICES
-          })
-        )
-      );
-      expect(featuredServices).toStrictEqual(MOCK_FEATURED_SERVICES);
-    });
-
-    it("should return undefined when not pot.some", () => {
-      expect(
-        featuredServicesSelector(appReducer(undefined, {} as Action))
-      ).toStrictEqual([]);
-
-      expect(
-        featuredServicesSelector(
-          appReducer({} as GlobalState, featuredServicesGet.request())
-        )
-      ).toStrictEqual([]);
-
-      expect(
-        featuredServicesSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredServicesGet.failure(MOCK_NETWORK_ERROR)
-          )
-        )
-      ).toStrictEqual([]);
-    });
-  });
-
-  describe("isLoadingFeaturedServicesSelector", () => {
-    it("should return true when pot.loading", () => {
-      const isLoading = isLoadingFeaturedServicesSelector(
-        appReducer({} as GlobalState, featuredServicesGet.request())
-      );
-      expect(isLoading).toStrictEqual(true);
-    });
-
-    it("should return false when not pot.loading", () => {
-      expect(
-        isLoadingFeaturedServicesSelector(appReducer(undefined, {} as Action))
-      ).toStrictEqual(false);
-
-      expect(
-        isLoadingFeaturedServicesSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredServicesGet.success({
-              services: MOCK_FEATURED_SERVICES
-            })
-          )
-        )
-      ).toStrictEqual(false);
-    });
-  });
-
-  describe("isErrorFeaturedServicesSelector", () => {
-    it("should return true when pot.error", () => {
-      const isError = isErrorFeaturedServicesSelector(
-        appReducer(
-          {} as GlobalState,
-          featuredServicesGet.failure(MOCK_NETWORK_ERROR)
-        )
-      );
-      expect(isError).toStrictEqual(true);
-    });
-
-    it("should return false when not pot.error", () => {
-      expect(
-        isErrorFeaturedServicesSelector(appReducer(undefined, {} as Action))
-      ).toStrictEqual(false);
-
-      expect(
-        isErrorFeaturedServicesSelector(
-          appReducer(
-            {} as GlobalState,
-            featuredServicesGet.success({
-              services: MOCK_FEATURED_SERVICES
-            })
-          )
-        )
-      ).toStrictEqual(false);
-    });
   });
 });
